Implement setprogname
Create an implementation of setprogname that sets the global __progname variable
Modify getprogname to use the new global __progname variable, but still generates its own name if __progname is NULL
Move progname functions to err.c
