--- util/private.h
@@ -1,6 +1,0 @@
-#ifndef LIBUTIL_PRIVATE_H
-#define LIBUTIL_PRIVATE_H
-
-const char* getprogname (void);
-
-#endif
